Changes to make this PR easier to read
