+++ Tests/AWSSDKSwiftCoreTests/Credential/ExpiringCredentialTests.swift
@@ -38,14 +38,14 @@ class ExpiringCredentialTests: XCTestCase {
         XCTAssert(!cred.isExpiring(within: expiringIn - 1))
         XCTAssert(!cred.isExpired)
         
-        /*let neverExpiring = RotatingCredential(
+        let neverExpiring = RotatingCredential(
             accessKeyId: accessKeyId,
             secretAccessKey: secretAccessKey,
             sessionToken: sessionToken,
             expiration: nil)
         
         XCTAssert(!neverExpiring.isExpiring(within: -5000))
-        XCTAssert(!neverExpiring.isExpired)*/
+        XCTAssert(!neverExpiring.isExpired)
     }
 }
